Simple exercise with React and React-Dom
Exercise with React and React-Dom that simulates the mechanics of a table for orders. The data comes from PrinceInput.js which acts as a database in this exercise. It is possible to modify the price and quantity field, so that the total price of the order changes.
Three Javascript libraries/framework are imported for this program:
- react
- react-dom
- parcel-bundler